fixing issue where there are files that match the statement are not symlinks

+++ modules/templates/preprocess_swgs.sh
@@ -47,8 +47,13 @@ fi
 if [[ -e "!{samples.externalSampleID}.bam" ]]
 then
 	for i in "!{samples.externalSampleID}."*bam*
-	do  
-		mv $(readlink ${i}) "!{samples.projectResultsDir}/alignment/"
+	do
+		if [[ -L "${i}" ]]
+		then
+			mv $(readlink ${i}) "!{samples.projectResultsDir}/alignment/"
+		else
+			rsync -v "${i}" "!{samples.projectResultsDir}/alignment/"
+		fi
 	done
 	rename "!{samples.combinedIdentifier}" "!{samples.externalSampleID}" "!{samples.projectResultsDir}/alignment/"*
 fi
